removed mvnw to be added to .gitignore and database fix
