+++ backend/src/main/java/com/MerchStore/backend/Dao/CartListDao.java
@@ -1,6 +1,6 @@
 package com.MerchStore.backend.Dao;
 
-import com.MerchStore.backend.ConnectionPooling.FlywayService.ConnectionManager;
+import com.MerchStore.backend.ConnectionPooling.ConnectionManager;
 import com.MerchStore.backend.Model.CartList;
 
 import java.sql.Connection;
@@ -22,8 +22,8 @@ public class CartListDao implements Dao<CartList>{
     public Optional<CartList> get(long id) {
         String statement = "SELECT * FROM cart_list where cart_id = ?";
 
+        Connection connection = ConnectionManager.getConnection();
         try{
-            Connection connection = ConnectionManager.getConnection();
             PreparedStatement preparedStatement = connection.prepareStatement(statement);
             preparedStatement.setLong(1, id);
             ResultSet resultSet = preparedStatement.executeQuery();
@@ -34,6 +34,8 @@ public class CartListDao implements Dao<CartList>{
         } catch (SQLException e) {
             System.out.println(e.getMessage());
             e.printStackTrace();
+        }finally {
+            ConnectionManager.releaseConnection(connection);
         }
         return Optional.empty();
     }
@@ -42,8 +44,8 @@ public class CartListDao implements Dao<CartList>{
     public List<CartList> getAll() {
         String statement = "SELECT * FROM cart_list where cart_id = ? ";
         LinkedList<CartList> resultList = new LinkedList<>();
+        Connection connection = ConnectionManager.getConnection();
         try{
-            Connection connection = ConnectionManager.getConnection();
             PreparedStatement preparedStatement = connection.prepareStatement(statement);
             preparedStatement.setLong(1,this.cart_id);
             ResultSet resultSet = preparedStatement.executeQuery();
@@ -54,6 +56,8 @@ public class CartListDao implements Dao<CartList>{
         } catch (SQLException e) {
             System.out.println(e.getMessage());
             e.printStackTrace();
+        }finally {
+            ConnectionManager.releaseConnection(connection);
         }
         return resultList;
     }
@@ -61,8 +65,8 @@ public class CartListDao implements Dao<CartList>{
     @Override
     public boolean save(CartList cart_list){
         String statement = "INSERT INTO cart_list values (?, ?, ?)";
+        Connection connection = ConnectionManager.getConnection();
         try{
-            Connection connection = ConnectionManager.getConnection();
             PreparedStatement preparedStatement = connection.prepareStatement(statement);
             preparedStatement.setLong(1, cart_list.getCartId());
             preparedStatement.setLong(2, cart_list.getProductId());
@@ -72,6 +76,8 @@ public class CartListDao implements Dao<CartList>{
         }catch (SQLException e){
             System.out.println(e.getMessage());
             e.printStackTrace();
+        }finally {
+            ConnectionManager.releaseConnection(connection);
         }
         return false;
     }
@@ -79,8 +85,8 @@ public class CartListDao implements Dao<CartList>{
     @Override
     public boolean update(CartList cart_list) {
         String statement = "UPDATE cart_list set product_quantity = ? where cart_id = ? AND product_id= ? ";
+        Connection connection = ConnectionManager.getConnection();
         try{
-            Connection connection = ConnectionManager.getConnection();
             PreparedStatement preparedStatement = connection.prepareStatement(statement);
             preparedStatement.setInt(1, cart_list.getQuantity());
             preparedStatement.setLong(2, cart_list.getCartId());
@@ -89,6 +95,8 @@ public class CartListDao implements Dao<CartList>{
         }catch (SQLException e){
             System.out.println(e.getMessage());
             e.printStackTrace();
+        }finally {
+            ConnectionManager.releaseConnection(connection);
         }
         return false;
     }
@@ -96,8 +104,8 @@ public class CartListDao implements Dao<CartList>{
     @Override
     public boolean delete(CartList cart_list) {
         String statement = "DELETE from cart_list where cart_id = ? AND product_id= ? ";
+        Connection connection = ConnectionManager.getConnection();
         try{
-            Connection connection = ConnectionManager.getConnection();
             PreparedStatement preparedStatement = connection.prepareStatement(statement);
             preparedStatement.setLong(1, cart_list.getCartId());
             preparedStatement.setLong(2, cart_list.getProductId());
@@ -105,6 +113,8 @@ public class CartListDao implements Dao<CartList>{
         } catch (SQLException e) {
             System.out.println(e.getMessage());
             e.printStackTrace();
+        }finally {
+            ConnectionManager.releaseConnection(connection);
         }
         return false;
     }
