+++ backend/src/main/java/com/MerchStore/backend/Dao/UsersDao.java
@@ -1,6 +1,6 @@
 package com.MerchStore.backend.Dao;
 
-import com.MerchStore.backend.ConnectionPooling.FlywayService.ConnectionManager;
+import com.MerchStore.backend.ConnectionPooling.ConnectionManager;
 import com.MerchStore.backend.Model.Users;
 import org.springframework.stereotype.Component;
 
@@ -19,8 +19,8 @@ public class UsersDao implements Dao<Users> {
     public Optional<Users> get(long id) {
         String statement = "SELECT * FROM users where user_id = ?";
 
+        Connection connection = ConnectionManager.getConnection();
         try {
-            Connection connection = ConnectionManager.getConnection();
             PreparedStatement preparedStatement = connection.prepareStatement(statement);
             preparedStatement.setLong(1, id);
             ResultSet resultSet = preparedStatement.executeQuery();
@@ -31,6 +31,8 @@ public class UsersDao implements Dao<Users> {
             }
         } catch (SQLException e) {
             System.out.println(e.getMessage());
+        }finally {
+            ConnectionManager.releaseConnection(connection);
         }
         return Optional.empty();
     }
@@ -39,8 +41,8 @@ public class UsersDao implements Dao<Users> {
     public List<Users> getAll() {
         String statement = "SELECT * FROM users";
         LinkedList<Users> resultList = new LinkedList<>();
+        Connection connection = ConnectionManager.getConnection();
         try {
-            Connection connection = ConnectionManager.getConnection();
             ResultSet resultSet = connection.prepareStatement(statement).executeQuery();
             while (resultSet.next()) {
                 Users user = new Users(resultSet.getLong("user_id"), resultSet.getString("first_name"), resultSet.getString("last_name"), resultSet.getString("email"), resultSet.getString("phone_number"));
@@ -48,6 +50,8 @@ public class UsersDao implements Dao<Users> {
             }
         } catch (SQLException e) {
             System.out.println(e.getMessage());
+        }finally {
+            ConnectionManager.releaseConnection(connection);
         }
         return resultList;
     }
@@ -55,8 +59,8 @@ public class UsersDao implements Dao<Users> {
     @Override
     public boolean save(Users user) {
         String statement = "INSERT INTO users values (?, ?, ?, ?, ?, ?)";
+        Connection connection = ConnectionManager.getConnection();
         try {
-            Connection connection = ConnectionManager.getConnection();
             PreparedStatement preparedStatement = connection.prepareStatement(statement);
             preparedStatement.setLong(1, user.getUserId());
             preparedStatement.setString(2, user.getFirstName());
@@ -68,6 +72,8 @@ public class UsersDao implements Dao<Users> {
             return preparedStatement.executeUpdate() == 1;
         } catch (SQLException e) {
             System.out.println(e.getMessage());
+        }finally {
+            ConnectionManager.releaseConnection(connection);
         }
         return false;
     }
@@ -75,8 +81,8 @@ public class UsersDao implements Dao<Users> {
     @Override
     public boolean update(Users user) {
         String statement = "UPDATE users set (first_name, last_name, email, phone_number, active) = (?, ?, ?, ?, ?) where user_id = ?";
+        Connection connection = ConnectionManager.getConnection();
         try {
-            Connection connection = ConnectionManager.getConnection();
             PreparedStatement preparedStatement = connection.prepareStatement(statement);
             preparedStatement.setString(1, user.getFirstName());
             preparedStatement.setString(2, user.getLastName());
@@ -88,6 +94,8 @@ public class UsersDao implements Dao<Users> {
             return preparedStatement.executeUpdate() == 1;
         } catch (SQLException e) {
             System.out.println(e.getMessage());
+        }finally {
+            ConnectionManager.releaseConnection(connection);
         }
         return false;
     }
@@ -95,14 +103,16 @@ public class UsersDao implements Dao<Users> {
     @Override
     public boolean delete(Users user) {
         String statement = "DELETE from users where user_id = ?";
+        Connection connection = ConnectionManager.getConnection();
         try {
-            Connection connection = ConnectionManager.getConnection();
             PreparedStatement preparedStatement = connection.prepareStatement(statement);
             preparedStatement.setLong(1, user.getUserId());
 
             return preparedStatement.executeUpdate() == 1;
         } catch (SQLException e) {
             System.out.println(e.getMessage());
+        }finally {
+            ConnectionManager.releaseConnection(connection);
         }
         return false;
     }
@@ -110,8 +120,8 @@ public class UsersDao implements Dao<Users> {
     public Optional<Users> getByEmail(String email) {
         String statement = "SELECT * FROM users where email = ?";
 
+        Connection connection = ConnectionManager.getConnection();
         try {
-            Connection connection = ConnectionManager.getConnection();
             PreparedStatement preparedStatement = connection.prepareStatement(statement);
             preparedStatement.setString(1, email);
             ResultSet resultSet = preparedStatement.executeQuery();
@@ -122,6 +132,8 @@ public class UsersDao implements Dao<Users> {
             }
         } catch (SQLException e) {
             System.out.println(e.getMessage());
+        }finally {
+            ConnectionManager.releaseConnection(connection);
         }
         return Optional.empty();
     }
