+++ backend/src/main/java/com/MerchStore/backend/Dao/VerifyUserDao.java
@@ -1,6 +1,6 @@
 package com.MerchStore.backend.Dao;
 
-import com.MerchStore.backend.ConnectionPooling.FlywayService.ConnectionManager;
+import com.MerchStore.backend.ConnectionPooling.ConnectionManager;
 import com.MerchStore.backend.Model.VerifyUser;
 import org.springframework.stereotype.Component;
 
@@ -17,8 +17,8 @@ public class VerifyUserDao implements Dao<VerifyUser>{
     public Optional<VerifyUser> get(long id) {
         String statement = "SELECT * FROM verify_user where user_id = ?";
 
+        Connection connection = ConnectionManager.getConnection();
         try {
-            Connection connection = ConnectionManager.getConnection();
             PreparedStatement preparedStatement = connection.prepareStatement(statement);
             preparedStatement.setLong(1, id);
             ResultSet resultSet = preparedStatement.executeQuery();
@@ -28,6 +28,8 @@ public class VerifyUserDao implements Dao<VerifyUser>{
             }
         } catch (SQLException e) {
             System.out.println(e.getMessage());
+        }finally {
+            ConnectionManager.releaseConnection(connection);
         }
         return Optional.empty();
     }
@@ -40,8 +42,8 @@ public class VerifyUserDao implements Dao<VerifyUser>{
     @Override
     public boolean save(VerifyUser verifyUser) {
         String statement = "INSERT INTO verify_user values (?, ?)";
+        Connection connection = ConnectionManager.getConnection();
         try {
-            Connection connection = ConnectionManager.getConnection();
             PreparedStatement preparedStatement = connection.prepareStatement(statement);
             preparedStatement.setLong(1, verifyUser.getUserId());
             preparedStatement.setString(2, verifyUser.getVerificationCode());
@@ -49,6 +51,8 @@ public class VerifyUserDao implements Dao<VerifyUser>{
             return preparedStatement.executeUpdate() == 1;
         } catch (SQLException e) {
             System.out.println(e.getMessage());
+        }finally {
+            ConnectionManager.releaseConnection(connection);
         }
         return false;
     }
@@ -56,8 +60,8 @@ public class VerifyUserDao implements Dao<VerifyUser>{
     @Override
     public boolean update(VerifyUser verifyUser) {
         String statement = "UPDATE verify_user set verification_code = ? where user_id = ?";
+        Connection connection = ConnectionManager.getConnection();
         try {
-            Connection connection = ConnectionManager.getConnection();
             PreparedStatement preparedStatement = connection.prepareStatement(statement);
             preparedStatement.setString(1, verifyUser.getVerificationCode());
             preparedStatement.setLong(2, verifyUser.getUserId());
@@ -65,6 +69,8 @@ public class VerifyUserDao implements Dao<VerifyUser>{
             return preparedStatement.executeUpdate() == 1;
         } catch (SQLException e) {
             System.out.println(e.getMessage());
+        }finally {
+            ConnectionManager.releaseConnection(connection);
         }
         return false;
     }
@@ -77,8 +83,8 @@ public class VerifyUserDao implements Dao<VerifyUser>{
     public Optional<VerifyUser> getByCode(String code) {
         String statement = "SELECT * FROM verify_user where verification_code = ?";
 
+        Connection connection = ConnectionManager.getConnection();
         try {
-            Connection connection = ConnectionManager.getConnection();
             PreparedStatement preparedStatement = connection.prepareStatement(statement);
             preparedStatement.setString(1, code);
             ResultSet resultSet = preparedStatement.executeQuery();
@@ -88,6 +94,8 @@ public class VerifyUserDao implements Dao<VerifyUser>{
             }
         } catch (SQLException e) {
             System.out.println(e.getMessage());
+        }finally {
+            ConnectionManager.releaseConnection(connection);
         }
         return Optional.empty();
     }
